Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
+#include <math.h>
 int main() {
     int a,b;char c;
     scanf("%d %d %c",&a,&b,&c);
-    float sum=(a+b);
-    float diff=(a-b);
-    float product=(a*b);
-    float div=(a/b);
+    int sum=(a+b);
+    int diff=(a-b);
+    int product=(a*b);
+    int div=(a/b);
     switch(c) {
         case '+' : printf("%d",sum);break;
         case '-' : printf("%d",diff);break;
